refactor code convention
