+++ .terraform.lock.hcl
@@ -1,0 +1,10 @@
+# This file is maintained automatically by "terraform init".
+# Manual edits may be lost in future updates.
+
+provider "test.com/hashicorp/ipcontrol" {
+  version     = "1.0.0"
+  constraints = ">= 1.0.0"
+  hashes = [
+    "h1:e2kDzVLIiswB6Tv939VijIDyx28pKmkD9L0d3Ezkae8=",
+  ]
+}
