+++ .vscode/launch.json
@@ -1,0 +1,24 @@
+{
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "name": "Launch Plugin",
+      "type": "go",
+      "request": "launch",
+      "mode": "auto",
+      "program": "${workspaceFolder}",
+      "args": [],
+      "env": {
+        "TF_LOG": "DEBUG", // Cho phép debug log cho Terraform
+        "TF_LOG_PATH": "${workspaceFolder}/terraform-debug.log" // Đường dẫn lưu log
+      }
+    },
+    {
+      "name": "Attach to Process",
+      "type": "go",
+      "request": "attach",
+      "mode": "local",
+      "processId": 2028
+  }
+  ]
+}
