+++ ipcontrol/models/ipc_subnet.go
@@ -2,7 +2,7 @@ package models
 
 import "terraform-provider-ipcontrol/ipcontrol/entities"
 
-func Subnet(subnet entities.IPC_Subnet) *entities.IPC_Subnet {
+func Subnet(subnet entities.IPCSubnet) *entities.IPCSubnet {
 	res := subnet
 	return &res
 }
